新增排班設定測試
